baremetal assets and applications plugins
