+++ frontend/src/routes/Governance/Governance.tsx
@@ -13,7 +13,7 @@ import { Link, Route, Switch, useLocation } from 'react-router-dom'
 import { useRecoilState } from 'recoil'
 import { acmRouteState, placementBindingsState, policiesState } from '../../atoms'
 import { NavigationPath } from '../../NavigationPath'
-import { PageContext } from '../Infrastructure/Clusters/Clusters'
+import { PageContext } from '../Infrastructure/Clusters/ClustersPage'
 import GovernanceOverview from './overview/Overview'
 import PoliciesPage from './policies/Policies'
 import PolicySetsPage from './policy-sets/PolicySets'
