+++ frontend/src/routes/Infrastructure/BareMetalAssets/BareMetalAssetsPage.test.tsx
